move test support to enzyme_utils, use async find wrapper in datamodel integraiton test
