+++ frontend/test/metabase/admin/datamodel/datamodel.e2e.spec.js
@@ -5,7 +5,13 @@ import {
   deleteAllSegments,
   deleteAllMetrics,
 } from "__support__/e2e_tests";
-import { click, clickButton, setInputValue } from "__support__/enzyme_utils";
+import {
+  click,
+  clickButton,
+  setInputValue,
+  enhanceEnzymeWrapper,
+} from "__support__/enzyme_utils";
+
 import { mount } from "enzyme";
 import { UPDATE_PREVIEW_SUMMARY } from "metabase/admin/datamodel/datamodel";
 import { FETCH_TABLE_METADATA } from "metabase/redux/metadata";
@@ -28,8 +34,6 @@ import {
   tables as Tables,
   fields as Fields,
 } from "metabase/entities";
-import { FETCH_DATABASE_IDFIELDS } from "metabase/entities/databases";
-import { push } from "react-router-redux";
 
 describe("admin/datamodel", () => {
   beforeAll(async () => useSharedAdminLogin());
@@ -39,33 +43,33 @@ describe("admin/datamodel", () => {
       const store = await createTestStore();
 
       store.pushPath("/admin/datamodel/database");
-      const app = mount(store.getAppContainer());
-
-      await store.waitForActions([Databases.actionTypes.FETCH_LIST]);
+      const app = enhanceEnzymeWrapper(mount(store.getAppContainer()));
+      await store.waitForActions([Tables.actions.fetchList]);
 
       // Open "Orders" table section
-      const adminListItems = app.find(".AdminList-item");
+      const adminListItems = await app.async.find(".AdminList-item");
       click(adminListItems.at(0));
-      await store.waitForActions([push("").type]);
+      await store.waitForActions([Tables.actions.fetchMetadata]);
 
       // Toggle its visibility to "Hidden"
-      click(app.find("#VisibilityTypes > span").at(1));
-      await store.waitForActions([Tables.actionTypes.UPDATE]);
+      const visibilityToggle = await app.async.find("#VisibilityTypes > span");
+      click(visibilityToggle.at(1));
+      await store.waitForActions([Tables.actions.update]);
 
       // Toggle "Why hide" to "Irrelevant/Cruft"
       click(app.find("#VisibilitySubTypes > span").at(2));
-      await store.waitForActions([Tables.actionTypes.UPDATE]);
+      await store.waitForActions([Tables.actions.update]);
 
       // Unhide
       click(app.find("#VisibilityTypes > span").at(0));
 
       // Open "People" table section
       click(adminListItems.at(1));
-      await store.waitForActions([push("").type]);
+      await store.waitForActions([Tables.actions.fetchMetadata]);
 
       // hide fields from people table
       // Set Address field to "Only in Detail Views"
-      const columnsListItems = app.find(ColumnsList).find("li");
+      const columnsListItems = (await app.async.find(ColumnsList)).find("li");
 
       click(columnsListItems.first().find(".TableEditor-field-visibility"));
       const onlyInDetailViewsRow = app
@@ -74,7 +78,7 @@ describe("admin/datamodel", () => {
         .at(1);
       expect(onlyInDetailViewsRow.text()).toMatch(/Only in Detail Views/);
       click(onlyInDetailViewsRow);
-      await store.waitForActions([Fields.actionTypes.UPDATE]);
+      await store.waitForActions([Fields.actions.update]);
 
       // Set Birth Date field to "Do Not Include"
       click(columnsListItems.at(1).find(".TableEditor-field-visibility"));
@@ -86,7 +90,7 @@ describe("admin/datamodel", () => {
       expect(doNotIncludeRow.text()).toMatch(/Do Not Include/);
       click(doNotIncludeRow);
 
-      await store.waitForActions([Fields.actionTypes.UPDATE]);
+      await store.waitForActions([Fields.actions.update]);
 
       // modify special type for address field
       click(columnsListItems.first().find(".TableEditor-field-special-type"));
@@ -96,7 +100,7 @@ describe("admin/datamodel", () => {
         .at(1);
       expect(entityNameTypeRow.text()).toMatch(/Entity Name/);
       click(entityNameTypeRow);
-      await store.waitForActions([Fields.actionTypes.UPDATE]);
+      await store.waitForActions([Fields.actions.update]);
 
       // TODO Atte Keinänen 8/9/17: Currently this test doesn't validate that the updates actually are reflected in QB
     });
@@ -109,8 +113,8 @@ describe("admin/datamodel", () => {
       const app = mount(store.getAppContainer());
 
       await store.waitForActions([
-        Databases.actionTypes.FETCH_LIST,
-        FETCH_DATABASE_IDFIELDS,
+        Databases.actions.fetchList,
+        Databases.actions.fetchIdfields,
       ]);
 
       // Click the new segment button and check that we get properly redirected
@@ -148,8 +152,8 @@ describe("admin/datamodel", () => {
       click(app.find('button[children="Save changes"]'));
 
       await store.waitForActions([
-        Segments.actionTypes.CREATE,
-        Databases.actionTypes.FETCH_LIST,
+        Segments.actions.create,
+        Databases.actions.fetchList,
       ]);
       expect(store.getPath()).toBe("/admin/datamodel/database/1/table/2");
 
@@ -171,8 +175,8 @@ describe("admin/datamodel", () => {
       const app = mount(store.getAppContainer());
 
       await store.waitForActions([
-        Databases.actionTypes.FETCH_LIST,
-        FETCH_DATABASE_IDFIELDS,
+        Databases.actions.fetchList,
+        Databases.actions.fetchIdfields,
       ]);
 
       // Click the new metric button and check that we get properly redirected
@@ -198,8 +202,8 @@ describe("admin/datamodel", () => {
       click(app.find('button[children="Save changes"]'));
 
       await store.waitForActions([
-        Metrics.actionTypes.CREATE,
-        Databases.actionTypes.FETCH_LIST,
+        Metrics.actions.create,
+        Databases.actions.fetchList,
       ]);
       expect(store.getPath()).toBe("/admin/datamodel/database/1/table/2");
