+++ frontend/test/metabase/admin/people/containers/EditUserModal.integ.spec.js
@@ -1,10 +1,7 @@
 import React from "react";
 import mock from "xhr-mock";
-import {
-  mountWithStore,
-  fillAndSubmitForm,
-  getFormValues,
-} from "__support__/integration_tests";
+import { mountWithStore } from "__support__/integration_tests";
+import { fillAndSubmitForm, getFormValues } from "__support__/enzyme_utils";
 
 import EditUserModal from "metabase/admin/people/containers/EditUserModal";
